src/schemata/schemata: update shemata
StatsSchema no longer has a "total" field. It is pretty obvious that the
fields within "weeks" are for weeks, and the fields not within "weeks"
are for total. This also makes yearly files more similar to the total
file.

Additionally, add a "years" field to TotalSchema to store the years that
the total file encompasses and make it easier to load data to the web.