src/web/src/static/index.js: char chart languages for total
Originally, individual languages were not selectable on the char chart
when total stats were selected due to performance concerns.

However, it feels wrong to limit users. If they have 10 years of
extraordinarily populated json files, they have put in the work, and
deserve to see this data, even if it takes a second or two to load.