src/services/config: remove deprecated settings
The refreshRate setting is no longer necessary as Coderank v0.4.0 will
not render buffer contents as it overcomplicates things and is useless.

The trackChars setting is deprecated because chars are the source of
a significant amount of data trends. If users truly wish for this
feature, it can easily be reincorporated in the future.