src/services/git: auto login with saved credentials
If the user has opted to save their credentials, the git service no
longer prompts them to enter their them. Instead, it uses their current
saved credentials to clone the repository on file, and if it fails,
grants the user the option to try again, prompting them to re-enter
their credentials in-case they have changed.